auto-tests/e2e_tests/requirements.txt
allure-pytest==2.14.2
allure-python-commons==2.14.2
annotated-types==0.7.0
attrs==25.3.0
iniconfig==2.1.0
packaging==25.0
pika==1.3.2
pluggy==1.5.0
psycopg==3.2.7
psycopg-binary==3.2.7
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pytest==8.3.5
python-dotenv==1.1.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2